Solution for day 24
It turns out the checks if a partitioning with the lowest 'quantum
entanglement' is not necessary with the given inputs. I left it in
anyway, in order to allow more general inputs.

Note that this still assumes that there exists at least one partitioning
for with the minimum length group (but that is a much weaker assumption
that it also exists for the lowest quantum entanglement group).